state: Check for lineage mismatches when writing
Accidentally losing a state with resources in it can be anywhere from
annoying to catastrophic. A common cause of such a problem is accidentally
clobbering a remote state with an unrelated local state or vice-versa.

Here we introduce safety checks that exploit the new "lineage" concept
to ensure that once a state location is established it can only be
updated with new states from the same lineage as the initial state.

We also make the remote state caching mechanism treat a lineage mismatch
as a confict, thus ensuring that automatic state syncing will stop if
the user manages to get into a broken, mismatched state.

As a special exception, we allow *empty* states to be "clobbered"
regardless of lineage. This enables remote state to be configured
easily in the common case where earlier user actions have implicitly
caused an empty local state to be created, and also allows the remote
state left behind from a config that has been destroyed to be overwritten
by a new, unrelated state from a different config.

The priority here is preventing actions that should never occur, so the
UX has not been polished. Later we may wish to make changes at higher
levels of abstraction to either prevent these situations from arising
in the first place (e.g. making remote configuration automatic based on
config) or giving the user more guidance on resolution.